[NA] Core 2.0-BreakingBad => Main RI
git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/ERM.Core/Main/BLCore;C15463

+++ Source/ApplicationServices/2Gis.Erm.BLCore.Aggregates/Activities/ReadModel/ActivitySpecs.cs
@@ -49,10 +49,10 @@ namespace DoubleGis.Erm.BLCore.Aggregates.Activities.ReadModel
                                              .Where(y => y.PropertyId == DealIdIdentity.Instance.Id)
                                              .Select(y => y.NumericValue)
                                              .FirstOrDefault(),
-                            Status = (ActivityStatus)x.ActivityPropertyInstances
-                                                      .Where(y => y.PropertyId == StatusIdentity.Instance.Id)
-                                                      .Select(y => y.NumericValue)
-                                                      .FirstOrDefault(),
+                            Status = (ActivityStatus)(int)x.ActivityPropertyInstances
+                                                           .Where(y => y.PropertyId == StatusIdentity.Instance.Id)
+                                                           .Select(y => y.NumericValue)
+                                                           .FirstOrDefault(),
                         });
                 }
